# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/emails/mark-all-read": {
            "post": {
                "description": "批量标记所有未读邮件为已读（仅本地状态），可选择指定账号或全部账号",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emails"
                ],
                "summary": "全部标记为已读",
                "parameters": [
                    {
                        "description": "标记请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/fusionmail_internal_dto_request.MarkAllAsReadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/emails/mark-read": {
            "post": {
                "description": "批量标记邮件为已读（仅本地状态）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emails"
                ],
                "summary": "标记邮件为已读",
                "parameters": [
                    {
                        "description": "邮件 ID 列表",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.MarkAsReadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/emails/mark-unread": {
            "post": {
                "description": "批量标记邮件为未读（仅本地状态）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emails"
                ],
                "summary": "标记邮件为未读",
                "parameters": [
                    {
                        "description": "邮件 ID 列表",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.MarkAsReadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/emails/permanent-delete": {
            "post": {
                "description": "批量永久删除回收站中的邮件（物理删除，不可恢复）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emails"
                ],
                "summary": "批量永久删除邮件",
                "parameters": [
                    {
                        "description": "邮件 ID 列表",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.BatchPermanentDeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/emails/search": {
            "get": {
                "description": "全文搜索邮件（主题、发件人、正文）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emails"
                ],
                "summary": "搜索邮件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "账户 UID",
                        "name": "account_uid",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码（默认 1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量（默认 20，最大 100）",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fusionmail_internal_service.EmailListResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/emails/stats/{account_uid}": {
            "get": {
                "description": "获取指定账户的邮件统计信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emails"
                ],
                "summary": "获取账户邮件统计",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账户 UID",
                        "name": "account_uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fusionmail_internal_service.AccountEmailStats"
                        }
                    }
                }
            }
        },
        "/api/v1/emails/unread-count": {
            "get": {
                "description": "获取指定账户或全部账户的未读邮件数",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emails"
                ],
                "summary": "获取未读邮件数",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账户 UID",
                        "name": "account_uid",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer",
                                "format": "int64"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/emails/{id}": {
            "get": {
                "description": "根据 ID 获取邮件的完整信息，包括附件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emails"
                ],
                "summary": "获取邮件详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "邮件 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除邮件（软删除，仅本地状态）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emails"
                ],
                "summary": "删除邮件",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "邮件 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/emails/{id}/archive": {
            "post": {
                "description": "归档邮件（仅本地状态）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emails"
                ],
                "summary": "归档邮件",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "邮件 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/emails/{id}/permanent": {
            "delete": {
                "description": "永久删除回收站中的邮件（物理删除，不可恢复）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emails"
                ],
                "summary": "永久删除邮件",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "邮件 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/emails/{id}/restore": {
            "post": {
                "description": "将邮件从本地垃圾箱恢复（仅本地状态）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emails"
                ],
                "summary": "恢复已删除邮件",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "邮件 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/emails/{id}/toggle-star": {
            "post": {
                "description": "切换邮件的星标状态（仅本地状态）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "emails"
                ],
                "summary": "切换星标状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "邮件 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/emails": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取邮件列表，支持分页、筛选和排序",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "邮件管理"
                ],
                "summary": "获取邮件列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "账户 UID",
                        "name": "account_uid",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否已读",
                        "name": "is_read",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否星标",
                        "name": "is_starred",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否归档",
                        "name": "is_archived",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "发件人地址（模糊匹配）",
                        "name": "from_address",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "主题（模糊匹配）",
                        "name": "subject",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "开始日期（YYYY-MM-DD）",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束日期（YYYY-MM-DD）",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码（默认 1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量（默认 20，最大 100）",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "fusionmail_internal_dto_request.MarkAllAsReadRequest": {
            "type": "object",
            "properties": {
                "account_uid": {
                    "description": "账户 UID，为空则应用于所有账号",
                    "type": "string"
                }
            }
        },
...
        "fusionmail_internal_service.AccountEmailStats": {
            "type": "object",
            "properties": {
                "archived_count": {
                    "type": "integer"
                },
                "starred_count": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                },
                "unread_count": {
                    "type": "integer"
                }
            }
        },
...
        "fusionmail_internal_service.EmailListResponse": {
            "type": "object",
            "properties": {
                "emails": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fusionmail_internal_service.EmailListItem"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "internal_handler.BatchPermanentDeleteRequest": {
            "type": "object",
            "required": [
                "ids"
            ],
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "internal_handler.MarkAsReadRequest": {
            "type": "object",
            "required": [
                "ids"
            ],
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}